vm_x64: Make HOTCOUNT table key less predictable
The original code makes a HOTCOUNT table key simply by taking the
lowest bits (except for bit 0 which is shifted away.) This has the
property that bytecodes within a prototype that are at a
HOTCOUNT_SIZE*2 offset from each other will always collide with each
other.

This is unfortunate because performance can be impacted by whether the
instructions share the same hash bucket, and this relationship can
change unpredictably when the source code is edited and the relative
offsets of the bytecode instructions changes.

This change simply does one additional bit manipulation to make the
relationship less direct. The bucket used to be (PC>>1)&HOTCOUNT and
now it is ((PC>>1)^PC)&HOTCOUNT. This takes one extra XOR instruction.